simplified code, ratio naalysis on unaligned images
